+++ projects/brazil_bu_chem/scripts/exhaust.R
@@ -120,6 +120,33 @@ for (i in seq_along(metadata$vehicles)) {
     }
     
     
+    
+    if(IM) {
+      #IM
+      ok <- im_ok[[metadata$vehicles[i]]][1:ncol(x)]
+      nok <- 1 - ok
+      
+      if(pol[j] == "CO") fim <- im_co[[metadata$vehicles[i]]][1:ncol(x)]
+      
+      if(pol[j] %in% c("HC", "NMHC")) fim <- im_hc[[metadata$vehicles[i]]][1:ncol(x)]
+      
+      if(pol[j] %in% c("NO", "NO2", "NOx")) fim <- im_nox[[metadata$vehicles[i]]][1:ncol(x)]
+      
+      if(pol[j] %in% c("PM")) fim <- im_pm[[metadata$vehicles[i]]][1:ncol(x)]
+      
+      # check that fim is only 1
+      
+      if(length(unique(fim)) == 1){
+        if(unique(fim) == 1){
+          ok <- 1                                
+        }
+      } 
+      
+      ef <- as.numeric(ef)*ok + as.numeric(ef)*fim*nok
+      ef[(length(ef)-8):length(ef)] <- max(ef[(length(ef)-8):length(ef)])
+      ef <- EmissionFactors(ef)
+    }
+    
     array_x <- emis(
       veh = x,
       lkm = lkm,